port cross check to a separate file
